+++ src/main/java/com/barabank/beans/Transaction.java
@@ -3,54 +3,60 @@ package com.barabank.beans;
 
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Date;
+
+/**
+ * @author Ilya Mikheev
+ *
+ */
 
 @Entity
 @Table(name = "transaction")
 public class Transaction implements Serializable {
 
-    private BigInteger id;
-    private BigInteger fromAccount;
-    private BigInteger toAccount;
-    private BigInteger sum;
+    private long id;
+    private long fromAccount;
+    private long toAccount;
+    private BigDecimal sum;
     private Date date;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    public BigInteger getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(long id) {
         this.id = id;
     }
 
     @Column(name = "from_account")
-    public BigInteger getFromAccount() {
+    public long getFromAccount() {
         return fromAccount;
     }
 
-    public void setFromAccount(BigInteger fromAccount) {
+    public void setFromAccount(long fromAccount) {
         this.fromAccount = fromAccount;
     }
 
     @Column(name = "to_account")
-    public BigInteger getToAccount() {
+    public long getToAccount() {
         return toAccount;
     }
 
-    public void setToAccount(BigInteger toAccount) {
+    public void setToAccount(long toAccount) {
         this.toAccount = toAccount;
     }
 
     @Column(name = "sum")
-    public BigInteger getSum() {
+    public BigDecimal getSum() {
         return sum;
     }
 
-    public void setSum(BigInteger sum) {
+    public void setSum(BigDecimal sum) {
         this.sum = sum;
     }
